Add a verbosity feature (#54)
Verbosity can now be set using the `rimseval.VERBOSITY` levels. If it is set to 0, then only useful warnings should be displayed to the user. This feature should help with debugging in the future, i.e., when a error / warning is given out while using the program in non-installer mode, it actually means something.